return specific subtypes of integer in constructor
